requirements.txt
alembic==1.17.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
black==25.11.0
blinker==1.9.0
cffi==2.0.0
click==8.3.0
contourpy==1.3.3
cycler==0.12.1
Flask==3.1.2
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.61.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.7
mypy_extensions==1.1.0
numpy==2.3.5
packaging==25.0
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
psycopg2-binary==2.9.11
pycparser==2.23
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytokens==0.3.0
six==1.17.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.3
WTForms==3.2.1
